requirements.txt
dwave-ocean-sdk>=3.3.0

matplotlib==3.3.4; python_version>'3.5'
matplotlib==3.0.3; python_version=='3.5'

descartes==1.1.0
pyshp==2.1.3
